+++ CHANGELOG.md
@@ -7,6 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [10.0.6] - 2025-07-30
+
+### Changed
+
+- Support older browsers (ES6) in IIFE builds
+- Improve face detection prompt timings
+
+## [10.0.5] - 2025-07-23
+
+### Changed
+
+- Fixed agent mode when face detection is in use
+- Responsive styling update
+- Improve face detection support on older mobile devices
+- Remove neutral expression checks
+- Update selfie capture text prompts
+
+## [10.0.4] - 2025-07-09
+
+### Changed
+
+- Styling update
+- Show loading message for face detection
+- Update big smile prompt to make it clearer
+
+## [10.0.3] - 2025-07-09
+
+### Changed
+
+- Fix selfie capture image composition
+- Mirror selfie capture image preview
+
 ## [10.0.2] - 2025-07-08
 
 ### Changed